Move all visual pieces to Skinning namespace
