+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/DrawableSlideFan.cs
@@ -1,4 +1,4 @@
-using osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides;
+using osu.Game.Rulesets.Sentakki.Skinning.Default.Slides;
 using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
