+++ osu.Game.Rulesets.Sentakki/Skinning/Default/DotPiece.cs
@@ -6,7 +6,7 @@ using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default
 {
     public class DotPiece : CompositeDrawable
     {
