+++ osu.Game.Rulesets.Sentakki/Skinning/Default/HoldBody.cs
@@ -7,7 +7,7 @@ using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default
 {
     public class HoldBody : CompositeDrawable
     {
