+++ osu.Game.Rulesets.Sentakki/Skinning/Default/RingPiece.cs
@@ -3,7 +3,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default
 {
     public class RingPiece : CircularContainer
     {
