+++ osu.Game.Rulesets.Sentakki/Skinning/Default/Slides/ISlideChevron.cs
@@ -1,4 +1,4 @@
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default.Slides
 {
     public interface ISlideChevron
     {
