+++ osu.Game.Rulesets.Sentakki/Skinning/Default/Slides/SlideTapPiece.cs
@@ -1,4 +1,4 @@
-﻿using osu.Framework.Allocation;
+using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -7,23 +7,36 @@ using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default.Slides
 {
-    public class TapPiece : CompositeDrawable
+    public class SlideTapPiece : CompositeDrawable
     {
         // This will be proxied, so a must.
         public override bool RemoveWhenNotAlive => false;
 
-        public TapPiece()
+        public readonly Container Stars;
+        public readonly StarPiece SecondStar;
+
+        public SlideTapPiece()
         {
+            Size = new Vector2(75);
+
             Anchor = Anchor.Centre;
             Origin = Anchor.Centre;
             Scale = new Vector2(0f);
             Position = new Vector2(0, -SentakkiPlayfield.NOTESTARTDISTANCE);
+
             InternalChildren = new Drawable[]
             {
-                new NoteRingPiece(),
-                new DotPiece(),
+                Stars = new Container(){
+                    RelativeSizeAxes = Axes.Both,
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Children = new Drawable[]{
+                        new StarPiece(),
+                        SecondStar = new StarPiece { Rotation = 36 }
+                    }
+                },
             };
         }
 
@@ -33,7 +46,10 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces
         private void load(DrawableHitObject drawableObject)
         {
             accentColour.BindTo(drawableObject.AccentColour);
-            accentColour.BindValueChanged(colour => Colour = colour.NewValue, true);
+            accentColour.BindValueChanged(colour =>
+            {
+                Stars.Colour = colour.NewValue;
+            }, true);
         }
     }
 }
