+++ osu.Game.Rulesets.Sentakki/Skinning/Default/TapPiece.cs
@@ -1,4 +1,4 @@
-using osu.Framework.Allocation;
+﻿using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -7,36 +7,23 @@ using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default
 {
-    public class SlideTapPiece : CompositeDrawable
+    public class TapPiece : CompositeDrawable
     {
         // This will be proxied, so a must.
         public override bool RemoveWhenNotAlive => false;
 
-        public readonly Container Stars;
-        public readonly StarPiece SecondStar;
-
-        public SlideTapPiece()
+        public TapPiece()
         {
-            Size = new Vector2(75);
-
             Anchor = Anchor.Centre;
             Origin = Anchor.Centre;
             Scale = new Vector2(0f);
             Position = new Vector2(0, -SentakkiPlayfield.NOTESTARTDISTANCE);
-
             InternalChildren = new Drawable[]
             {
-                Stars = new Container(){
-                    RelativeSizeAxes = Axes.Both,
-                    Anchor = Anchor.Centre,
-                    Origin = Anchor.Centre,
-                    Children = new Drawable[]{
-                        new StarPiece(),
-                        SecondStar = new StarPiece { Rotation = 36 }
-                    }
-                },
+                new NoteRingPiece(),
+                new DotPiece(),
             };
         }
 
@@ -46,10 +33,7 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Slides
         private void load(DrawableHitObject drawableObject)
         {
             accentColour.BindTo(drawableObject.AccentColour);
-            accentColour.BindValueChanged(colour =>
-            {
-                Stars.Colour = colour.NewValue;
-            }, true);
+            accentColour.BindValueChanged(colour => Colour = colour.NewValue, true);
         }
     }
 }
