+++ osu.Game.Rulesets.Sentakki/Skinning/Default/TouchHolds/TouchHoldBody.cs
@@ -6,7 +6,7 @@ using osu.Game.Rulesets.Objects.Drawables;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.TouchHolds
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default.TouchHolds
 {
     public class TouchHoldBody : CircularContainer
     {
