+++ osu.Game.Rulesets.Sentakki/Skinning/Default/TouchHolds/TouchHoldProgressPiece.cs
@@ -1,30 +1,42 @@
+using System;
+using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Effects;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.TouchHolds
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default.TouchHolds
 {
-    public class TouchHoldCentrePiece : CompositeDrawable
+    public class TouchHoldProgressPiece : CompositeDrawable
     {
-        private OsuColour colours = new OsuColour();
-        public TouchHoldCentrePiece()
+        private readonly CircularProgress redProgress;
+        private readonly CircularProgress yellowProgress;
+        private readonly CircularProgress greenProgress;
+        private readonly CircularProgress blueProgress;
+
+        public BindableDouble ProgressBindable = new BindableDouble();
+
+        public TouchHoldProgressPiece()
         {
+            OsuColour colours = new OsuColour();
+            ProgressBindable.BindValueChanged(p =>
+            {
+                redProgress.Current.Value = Math.Min(p.NewValue, .25);
+                yellowProgress.Current.Value = Math.Min(p.NewValue, .50);
+                greenProgress.Current.Value = Math.Min(p.NewValue, .75);
+                blueProgress.Current.Value = p.NewValue;
+            });
             Origin = Anchor.Centre;
             Anchor = Anchor.Centre;
-            Size = new Vector2(80);
             Masking = true;
-            CornerRadius = 20f;
+            BorderColour = Color4.White;
+            BorderThickness = 3;
+            Alpha = .8f;
+            Size = new Vector2(110);
+            CornerRadius = 27.5f;
             Rotation = 45;
-            EdgeEffect = new EdgeEffectParameters
-            {
-                Type = EdgeEffectType.Shadow,
-                Colour = Color4.Black,
-                Radius = 10f,
-            };
             InternalChildren = new Drawable[]{
                 new Container {
                     Origin = Anchor.Centre,
@@ -33,45 +45,44 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.TouchHolds
                     Size = new Vector2(2),
                     Rotation = -45f,
                     Children = new Drawable[]{
-                        new CircularProgress{
+                        blueProgress = new CircularProgress{
                             Anchor = Anchor.Centre,
                             Origin = Anchor.Centre,
                             InnerRadius = 1,
                             Size = Vector2.One,
                             RelativeSizeAxes = Axes.Both,
-                            Current = { Value = 1 },
+                            Current = { Value = 0 },
                             Colour = colours.Blue
                         },
-                        new CircularProgress{
+                        greenProgress = new CircularProgress{
                             Anchor = Anchor.Centre,
                             Origin = Anchor.Centre,
                             InnerRadius = 1,
                             Size = Vector2.One,
                             RelativeSizeAxes = Axes.Both,
-                            Current = { Value = .75 },
+                            Current = { Value = 0 },
                             Colour = colours.Green
                         },
-                        new CircularProgress{
+                        yellowProgress = new CircularProgress{
                             Anchor = Anchor.Centre,
                             Origin = Anchor.Centre,
                             InnerRadius = 1,
                             Size = Vector2.One,
                             RelativeSizeAxes = Axes.Both,
-                            Current = { Value = .5 },
+                            Current = { Value = 0 },
                             Colour = colours.Yellow,
                         },
-                        new CircularProgress{
+                        redProgress = new CircularProgress{
                             Anchor = Anchor.Centre,
                             Origin = Anchor.Centre,
                             InnerRadius = 1,
                             Size = Vector2.One,
                             RelativeSizeAxes = Axes.Both,
-                            Current = { Value = .25 },
+                            Current = { Value = 0 },
                             Colour = colours.Red,
                         },
                     }
-                },
-                new DotPiece()
+                }
             };
         }
     }
