+++ osu.Game.Rulesets.Sentakki/Skinning/Default/Touches/TouchPiece.cs
@@ -5,7 +5,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osuTK;
 
-namespace osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces.Touches
+namespace osu.Game.Rulesets.Sentakki.Skinning.Default.Touches
 {
     public class TouchPiece : CompositeDrawable
     {
